[feat] createScene 구현

+++ core-model/src/main/kotlin/com/ddd/filmo/model/Scene.kt
@@ -8,7 +8,7 @@ data class Scene(
     val documentId: String? = "",
     val sceneText: String? = "",
     val sceneType: SceneType? = SceneType.fromColor(0xff000000),
-    val sceneRate: Int? = null,
+    val sceneRate: Float? = null,
     val movie: Movie? = null,
     val createdAt: Date = Date(),
 ):DomainModel {
@@ -21,7 +21,7 @@ data class Scene(
             documentId = "1",
             sceneText = "다람쥐 헌 쳇바퀴에 타고파 다람쥐 헌 쳇바퀴에 타고파 다람쥐 헌 쳇바퀴에 타고파 다람쥐 헌 쳇바퀴에 타고파 다람쥐 헌 쳇바퀴에 타고파 다람쥐 헌 쳇바퀴에 타고파 다람쥐 헌 쳇바퀴에 타고파 다람쥐 헌 쳇바퀴에 타고파 다람쥐 헌 쳇바퀴에 타고파 다람쥐 헌 쳇바퀴에 타고파 ",
             sceneType = SceneType.fromUrl("https://hips.hearstapps.com/hmg-prod/images/notebook-1551291989.jpg?crop=0.666xw:1.00xh;0.173xw,0&resize=1200:*"),
-            sceneRate = 5,
+            sceneRate = 5f,
             movie = Movie.mock,
             createdAt = Date(),
         )
@@ -30,7 +30,7 @@ data class Scene(
             documentId = "2",
             sceneText = "다람쥐 헌 쳇바퀴에 타고파 다람쥐 헌 쳇바퀴에 타고파 다람쥐 헌 쳇바퀴에 타고파 다람쥐 헌 쳇바퀴에 타고파 다람쥐 헌 쳇바퀴에 타고파 다람쥐 헌 쳇바퀴에 타고파 다람쥐 헌 쳇바퀴에 타고파 다람쥐 헌 쳇바퀴에 타고파 다람쥐 헌 쳇바퀴에 타고파 다람쥐 헌 쳇바퀴에 타고파 ",
             sceneType = SceneType.fromUrl("https://image.cine21.com/resize/cine21/still/2017/0304/18_07_58__58ba83ee3e0d3[W680-].jpg"),
-            sceneRate = 5,
+            sceneRate = 5f,
             movie = Movie.mock1,
             createdAt = Date(),
         )
@@ -39,7 +39,7 @@ data class Scene(
             documentId = "2",
             sceneText = "When a beautiful stranger leads computer hacker Neo to a forbidding underworld, he discovers the shocking truth--the life he knows is the elaborate deception of an evil cyber-intelligence.",
             sceneType = SceneType.fromUrl("https://image.cine21.com/resize/cine21/still/2017/0304/18_07_58__58ba83ee3e0d3[W680-].jpg"),
-            sceneRate = 5,
+            sceneRate = 5f,
             movie = Movie.mock1,
             createdAt = Date(),
         )
